+++ tests/debugger/oegdb/enc/enc.c
@@ -33,6 +33,22 @@ int enc_add(int a, int b)
     return c;
 }
 
+static void _raise_exceptions(void)
+{
+    // Raise CPUID. This is handled by oecore.
+    unsigned int c;
+    asm("movl  $1, %%eax   \n\t"
+        "cpuid             \n\t"
+        : "=c"(c)
+        :
+        : "eax", "ebx", "edx");
+
+    // Raise rdtsc. This is handled via a handler below.
+    register uint64_t rax __asm__("rax");
+    register uint64_t rdx __asm__("rdx");
+    asm volatile("rdtsc" : "=r"(rax), "=r"(rdx));
+}
+
 // The following function is intended to be called by the debugger.
 // It must be retained via OE_EXPORT.
 OE_EXPORT
@@ -44,6 +60,9 @@ int square(int x)
         oe_is_within_enclave(&x, sizeof(x)) ? "within enclave"
                                             : "outside enclave");
     g_square_called = true;
+
+    // Raising signal in a function called by gdb crashes gdb 8.0 and above.
+    // _raise_exceptions();
     return x * x;
 }
 
@@ -54,11 +73,34 @@ static void enclave_function(void)
     // The following assertion will fail if the debugger was not able to walk
     // the ocall stack back to the enclave and set the value of enc_magic.
     OE_TEST(enc_magic == MAGIC_VALUE);
+    _raise_exceptions();
 }
 
 void enc_test_stack_stitching(void)
 {
+    _raise_exceptions();
     enclave_function();
+}
+
+#define RDTSC_OPCODE 0x310f
+// 2nd-chance exception handler to continue on test triggered exceptions
+static uint64_t _rdtsc_sigill_handler(oe_exception_record_t* exception)
+{
+    if (exception->code == OE_EXCEPTION_ILLEGAL_INSTRUCTION)
+    {
+        if (*((uint16_t*)exception->context->rip) == RDTSC_OPCODE)
+        {
+            exception->context->rip += 2;
+            return OE_EXCEPTION_CONTINUE_EXECUTION;
+        }
+    }
+    return OE_EXCEPTION_CONTINUE_SEARCH;
+}
+
+__attribute__((constructor)) void enc_constructor(void)
+{
+    oe_add_vectored_exception_handler(false, _rdtsc_sigill_handler);
+    _raise_exceptions();
 }
 
 OE_SET_ENCLAVE_SGX(